+++ client/actions/actionTypes.js
@@ -9,12 +9,6 @@ export const SPOTIFY_TOP_ARTISTS_SUCCESS = 'SPOTIFY_TOP_ARTISTS_SUCCESS';
 export const SPOTIFY_TOP_TRACKS_BEGIN = 'SPOTIFY_TOP_TRACKS_BEGIN';
 export const SPOTIFY_TOP_TRACKS_SUCCESS = 'SPOTIFY_TOP_TRACKS_SUCCESS';
 
-export const SPOTIFY_FETCH_ARTIST_BEGIN = 'SPOTIFY_FETCH_ARTIST_BEGIN';
-export const SPOTIFY_FETCH_ARTIST_SUCCESS = 'SPOTIFY_FETCH_ARTIST_SUCCESS';
-export const SPOTIFY_FETCH_ARTIST_FAILURE = 'SPOTIFY_FETCH_ARTIST_FAILURE';
-export const SPOTIFY_FETCH_ARTIST_OPEN = 'SPOTIFY_FETCH_ARTIST_OPEN';
-export const SPOTIFY_FETCH_ARTIST_CLOSE = 'SPOTIFY_FETCH_ARTIST_CLOSE';
-
 export const SPOTIFY_CURRENT_TRACK_BEGIN = 'SPOTIFY_CURRENT_TRACK_BEGIN';
 export const SPOTIFY_CURRENT_TRACK_SUCCESS = 'SPOTIFY_CURRENT_TRACK_SUCCESS';
 export const SPOTIFY_CURRENT_TRACK_NOT_PLAYING =
@@ -22,3 +16,9 @@ export const SPOTIFY_CURRENT_TRACK_NOT_PLAYING =
 
 export const SPOTIFY_RECENT_TRACKS_BEGIN = 'SPOTIFY_RECENT_TRACKS_BEGIN';
 export const SPOTIFY_RECENT_TRACKS_SUCCESS = 'SPOTIFY_RECENT_TRACKS_SUCCESS';
+
+export const FETCH_ARTIST_BEGIN = 'FETCH_ARTIST_BEGIN';
+export const FETCH_ARTIST_SUCCESS = 'FETCH_ARTIST_SUCCESS';
+export const FETCH_ARTIST_FAILURE = 'FETCH_ARTIST_FAILURE';
+export const FETCH_ARTIST_OPEN = 'FETCH_ARTIST_OPEN';
+export const FETCH_ARTIST_CLOSE = 'FETCH_ARTIST_CLOSE';
