+++ client/components/App/index.js
@@ -31,7 +31,9 @@ export default function App() {
     enter: { opacity: 1, transform: 'translate3d(0, 0%,0)' },
     leave: { opacity: 0, transform: 'translate3d(0, 100%,0)' },
   });
-  const isArtistLoading = useSelector((state) => state.isArtistLoading);
+  const isArtistLoading = useSelector(
+    (state) => state.artistInfo.isArtistLoading
+  );
 
   return (
     <>
