+++ client/configureStore.js
@@ -1,5 +1,4 @@
 import { configureStore } from '@reduxjs/toolkit';
-//import rootReducer from './reducers';
 import reducer from './reducers';
 
 const store = configureStore({
